requirements.txt
python-telegram-bot
python-valve
python-dotenv